refactor(NcEmptyContent): migrate component to Typescript
