+++ src/components/NcAppSidebar/NcAppSidebar.vue
@@ -729,7 +729,7 @@ import IconStarOutline from 'vue-material-design-icons/StarOutline.vue'
 import NcAppSidebarTabs from './NcAppSidebarTabs.vue'
 import NcActions from '../NcActions/index.js'
 import NcButton from '../NcButton/index.ts'
-import NcEmptyContent from '../NcEmptyContent/index.js'
+import NcEmptyContent from '../NcEmptyContent/index.ts'
 import NcLoadingIcon from '../NcLoadingIcon/index.js'
 import NcAppSidebarHeader from '../NcAppSidebarHeader/index.ts'
 import Focus from '../../directives/Focus/index.ts'
