+++ src/components/NcEmptyContent/index.ts
