+++ src/components/NcRichText/NcReferencePicker/NcProviderList.vue
@@ -44,7 +44,7 @@
 <script>
 import { searchProvider } from './../../../functions/reference/providerHelper.js'
 import { isUrl } from './utils.js'
-import NcEmptyContent from '../../NcEmptyContent/index.js'
+import NcEmptyContent from '../../NcEmptyContent/index.ts'
 import NcHighlight from '../../NcHighlight/index.js'
 import NcSelect from '../../NcSelect/index.js'
 import { t } from '../../../l10n.js'
