+++ src/components/index.ts
@@ -53,7 +53,7 @@ export { default as NcDialog } from './NcDialog/index.js'
 export { default as NcDialogButton } from './NcDialogButton/index.js'
 export { default as NcEllipsisedOption } from './NcEllipsisedOption/index.js'
 export { default as NcEmojiPicker } from './NcEmojiPicker/index.js'
-export { default as NcEmptyContent } from './NcEmptyContent/index.js'
+export { default as NcEmptyContent } from './NcEmptyContent/index.ts'
 export { default as NcGuestContent } from './NcGuestContent/index.js'
 export { default as NcHeaderButton } from './NcHeaderButton/index'
 export { default as NcHeaderMenu } from './NcHeaderMenu/index.js'
